# src/main/resources/api/openapi.yaml
openapi: 3.1.0
info:
  title: Courtside
  version: 0.1.0
  summary: Court booking for a single sports club.
  description: |
    Courtside is single-tenant: this document describes the API of one club's own instance.
    There is no tenant identifier anywhere in it, and there never will be.

    This document is the source of truth for the API, not a description generated from an
    implementation. An operation that is not here is not part of the API, and a change here is a
    change to a published contract.

    ## Errors

    Every failure is an RFC 9457 problem detail. `type` is always a `urn:courtside:error:<slug>`
    URN: a stable identifier, not a fetchable page. A club runs its instance under its own domain,
    so an `https://` type would name a host nobody operates.

    Two arrays carry everything a client has to show a member, and both hold the same kind of
    entry:

    - `violations` — a rule the request broke. Always an array, even where only one entry is ever
      possible, so a client has one shape to render rather than two.
    - `fieldErrors` — the same, plus the name of the input it came from.

    Neither carries rendered text. Each entry has a `code` naming a message and `params` holding
    the values that message interpolates, and the client resolves the pair against its own message
    catalogue. This is what makes the API usable in a language the server does not ship.
  license:
    name: AGPL-3.0-or-later
    identifier: AGPL-3.0-or-later
servers:
...
security:
  - sessionCookie: []
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: SESSION
      description: |
        A server-side session, established by `POST /api/session`. Deliberately not a JWT: the
        club's own frontend and the API share an origin, and an admin must be able to end a
        session immediately, which a self-contained token cannot offer.
    csrfToken:
      type: apiKey
      in: header
      name: X-XSRF-TOKEN
      description: |
        Required on every state-changing request. Any `GET` issues a readable `XSRF-TOKEN` cookie;
        echo its value in this header. Modelled as a security scheme rather than as a header
        parameter because it is consumed before the operation runs and is never part of its input.

        A write without it answers `401`, not `403`: for an unauthenticated caller the missing
        token is indistinguishable from a missing session.
  schemas:
...
    RuleType:
      type: string
      enum: [OPENING_HOURS, SLOT_GRID, ADVANCE_WINDOW, MAX_OPEN_BOOKINGS]
...
    Violation:
      type: object
      description: |
        A rule the request broke, named rather than rendered. The client resolves `code` against
        its own message catalogue and interpolates `params`.
      required: [code, params]
      properties:
        code:
          type: string
          examples: ["booking.rule.advanceWindow.exceeded"]
        params:
          type: object
          additionalProperties: true
          examples: [{ maxDays: 14 }]
    FieldError:
      type: object
      description: A violation that also names the input it came from.
      required: [field, code, params]
      properties:
        field:
          type: string
          examples: ["endsAt"]
        code:
          type: string
          examples: ["validation.ChronologicalSlot"]
        params:
          type: object
          additionalProperties: true
    Problem:
      type: object
      description: RFC 9457 problem detail.
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          description: A `urn:courtside:error:<slug>` URN. Stable, and not a fetchable page.
          examples: ["urn:courtside:error:court-not-bookable"]
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
          description: |
            English, for a developer reading a log. Never shown to a member, and never built from
            a caller's input — the translatable part is in `violations`.
        instance:
          type: string
          format: uri
        violations:
          type: array
          items: { $ref: "#/components/schemas/Violation" }
        fieldErrors:
          type: array
          items: { $ref: "#/components/schemas/FieldError" }
...
    ActiveRequest:
      type: object
      required: [active]
      properties:
        active: { type: boolean }
...
    RuleSet:
      type: object
      required: [id, name, active]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        active: { type: boolean }
    RuleSetRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, maxLength: 60, pattern: ".*\\S.*" }
    RuleDefinition:
      type: object
      required: [ruleType, params]
      properties:
        ruleType: { $ref: "#/components/schemas/RuleType" }
        params:
          type: object
          additionalProperties: { type: integer }
    SetRuleRequest:
      type: object
      required: [params]
      properties:
        params:
          type: object
          additionalProperties: { type: integer }
...
  responses:
    Unauthenticated:
      description: No session, or a state-changing request without a CSRF token.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Forbidden:
      description: Authenticated, but not allowed to do this.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    NotFound:
      description: No such thing.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Invalid:
      description: The request was understood and rejected.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
paths:
...
  /api/admin/rule-sets:
    get:
      tags: [admin-rule-sets]
      operationId: listRuleSets
      summary: Every ruleset, active or not.
      responses:
        "200":
          description: The list.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/RuleSet" }
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [admin-rule-sets]
      operationId: createRuleSet
      summary: Create a ruleset.
      security:
        - sessionCookie: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RuleSetRequest" }
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RuleSet" }
        "400":
          $ref: "#/components/responses/Invalid"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: A rule set already carries that name.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /api/admin/rule-sets/{id}:
    get:
      tags: [admin-rule-sets]
      operationId: getRuleSet
      summary: One ruleset.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: The ruleset.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RuleSet" }
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [admin-rule-sets]
      operationId: changeRuleSet
      summary: Change a ruleset.
      security:
        - sessionCookie: []
          csrfToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RuleSetRequest" }
      responses:
        "200":
          description: Changed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RuleSet" }
        "400":
          $ref: "#/components/responses/Invalid"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: A rule set already carries that name.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /api/admin/rule-sets/{id}/active:
    put:
      tags: [admin-rule-sets]
      operationId: setRuleSetActive
      summary: Activate or deactivate a ruleset.
      description: An inactive rule set cannot be assigned to a membership type.
      security:
        - sessionCookie: []
          csrfToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ActiveRequest" }
      responses:
        "200":
          description: Changed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RuleSet" }
        "400":
          $ref: "#/components/responses/Invalid"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/admin/rule-sets/{id}/rules:
    get:
      tags: [admin-rule-sets]
      operationId: listRules
      summary: The rules configured in a rule set.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: One entry per configured rule type.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/RuleDefinition" }
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/admin/rule-sets/{id}/rules/{ruleType}:
    put:
      tags: [admin-rule-sets]
      operationId: setRule
      summary: Configure one rule of a rule set.
      description: |
        Rules are data, not code: a rule is a row with parameters, and which parameters a type
        accepts is defined by the type. An unknown parameter is refused rather than ignored.
      security:
        - sessionCookie: []
          csrfToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: ruleType
          in: path
          required: true
          schema: { $ref: "#/components/schemas/RuleType" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SetRuleRequest" }
      responses:
        "200":
          description: Configured.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RuleDefinition" }
        "400":
          $ref: "#/components/responses/Invalid"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [admin-rule-sets]
      operationId: removeRule
      summary: Remove a rule from a rule set.
      security:
        - sessionCookie: []
          csrfToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: ruleType
          in: path
          required: true
          schema: { $ref: "#/components/schemas/RuleType" }
      responses:
        "204":
          description: Removed.
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...